Consolidated KrogerAPI Auth Token functions into AuthTokenFactoryBean utility class.
